adjust cppcheck and fix more warnings

+++ include/qalgorithms_read_file.h
@@ -138,8 +138,8 @@ namespace qAlgorithms
 
     std::vector<char> encode_base64_dbl(const double *input_dbl, const size_t in_len_dbl);
 
-    int bytesToFloatVec(const std::vector<char> *bytes, const bool isDouble,
-                        std::vector<float> *result);
+    size_t bytesToFloatVec(const std::vector<char> *bytes, const bool isDouble,
+                           std::vector<float> *result);
 
     void compress_zlib(const std::vector<char> *uncompressed_string, std::vector<char> *output_string);
